# backend/docs/swagger.yaml
basePath: /api
definitions:
  handlers.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  handlers.UpdateOrderPaymentRequest:
    properties:
      payment_status:
        type: string
    type: object
  handlers.UpdateOrderStatusRequest:
    properties:
      status:
        type: string
    type: object
info:
  contact: {}
  description: Backend API for fish marketplace and aquaculture management.
  title: Fish Market & Aquaculture API
  version: "1.0"
paths:
...
  /orders:
    get:
      description: Get paginated orders for admin or current customer
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get orders
      tags:
      - Order
  /orders/{id}:
    get:
      description: Get order detail by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get order detail
      tags:
      - Order
  /orders/{id}/invoice:
    get:
      description: Generate and download order invoice PDF
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Download order invoice
      tags:
      - Order
  /orders/{id}/payment:
    put:
      consumes:
      - application/json
      description: Update order payment status
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payment status payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateOrderPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update payment status
      tags:
      - Order
  /orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Update order lifecycle status
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order status payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - Order
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"